Implement all undocumented opcodes (that make sense since some are effectively random)
Fix RTI so it pops status with PLP which then makes sure S1 and B are correct in P
Fix indirect JMP across a page boundary to actually work

Testing updates:

Add the ability to load the test log and verify state on each instruction
for the NES functional test. Make it pass based on the above fixes and also
don't hardcode the CPU type and actually use the one from the test or else the
NES version tries to do BCD when it shouldn't and correctly fails tests.