Clear up a few TODO's:
Implement Reset correctly as a 6 tick version and fixed tests to reflect.
Drop one TODO that didn't apply (B bit on PHP).

Fix the state capturing in the circular buffer during tests as we need
the RAM values for the instruction at the time of it running as some
of the tests are self modifying code..