Large scale refactoring to get a common Tick interface among components.
No actual interface but if needed one could be laid on top now. Tick just
returns an error. For the CPU portion add an InstructionDone method for
getting the instruction state.

Add TickDone to all chips so latching activity that takes place on the rising
edge of a clock gets dealt with. Mostly so the TIA can present a consistent
hPOS view for latching no matter the order of clock ticking for a given clock.