Update frame drawing so we can callback on specific x,y coordinates as well as just at start of line.
Rename TestPlayfield to TestDrawing as with the above it can turn it
into a general test bed for any of the drawing we want to verify
(already have the ability to write specific canonical images completely).

Add a test for basic error states in setup/etc.

Abstract setup to it's own routine for convenience.