Implement all real state changes in TickDone() so that ordering doesn't matter and Read calls always see a consistent view.
Checking the timing diagram examples in the 6532 datasheet shows
all of the signals are clocked, including interrupt states so this
lines up with reality. It does mean resetting the timer value on
the same clock it's going 00->FF will lose an interrupt but that's
expected.

Adjust the tests since all Read/Write calls need a Tick sequence
done after them to reflect reality of how this really works.