Fixes for player rendering:
In reset just make sure we set the right clock. The state machine
forces it to skip the first instance.

Accidentally hard coded checking player count/width in the advance
function so pass that in as a param instead.

Fixes for double/quad rendering where clocks were getting ticked wrong.