Implement playfield priority bits.
Found a bug in RSYNC tests where this wasn't accounting for ball vs PF
correctly.

In a few other tests either purposely turn on PFP to test or disable
the playfield in that test as it's just getting in the way (i.e
other tests prove things).

Write some tests for vertical delay working for the player sprite.