Sync up set of fixes that have been standing for a long time:
Add basic suport for paddles. Right now you have to init a VCS
with either joysticks or paddles but not both, etc. Long term
need a way to dynamically add/remove controllers.

Add debug support to all chips to track state when things run.
Turns out running a whole cart is easier to debug then.

Add a total clocks since start to each chip to track state since
some run at different speeds.

Fix the PIA impl which only ticked the timer to 0 for a single clock
which isn't how the chip works. It holds at 0 for an entire multiplier
cycle allowing the CPU to poll and find the 0 crossing. Otherwise
the CPU polling has to get lucky and for atari carts means frame jitter.

Need to fix the unit tests for the above since they don't account for
this yet.