Lots of work to make Digdug work.
This is the first cart with a SuperChip which requires not only mapping
RAM into somewhere in the ROM space (dual ported since there's no R/W line
to the cart), but also retaining bus state. If the write port is read
the last item on the databus will be the value written into it. This requires
that all memory.Bank implementations have the ability to see up to their Parent
holding memory.Bank (if implemented). Then a helper func in memory.go can walk
any given Bank to the top to show the whole system databus. Any implementation
of memory.Bank is now required to cache what it sees on the databus lines but
only the top level one actually has a full view (since every chip isn't active
on each clock).

To assist with this implement a basic RAM in memory.go so that carts and the PIA
don't have to write their own implemntations.